+++ test/activity/Activity-test.js
@@ -8,6 +8,10 @@ import {TaskBehaviour} from '../../src/tasks/Task';
 import {SignalTaskBehaviour} from '../../src/tasks/SignalTask';
 
 describe('Activity', () => {
+  describe('properties', () => {
+
+  });
+
   describe('run on inbound', () => {
     it('starts run when inbound sequence flow is taken', () => {
       const sequenceFlows = [];